Move version functions to SystemUtils.h
Create a new header file SystemUtils.h which will contain different
utilities related to the current system.

For now, SystemUtils.h contains the functions required to obtain,
parse and compare the system's Build ID version.
